Add function that creates an iterator
